% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-regex-utils.R
\name{glob_to_regex}
\alias{glob_to_regex}
\title{Change glob pattern to regular expression pattern}
\usage{
glob_to_regex(pattern)
}
\arguments{
\item{pattern}{A glob-style character vector}
}
\value{
A character vector containing regular expressions instead of
glob-style wildcards
}
\description{
Change glob wildcards to regular expressions. Word boundaries
are added if no wildcard was present. Thus, "dog" becomes "\\\\bdog\\\\b" and
"dog*" becomes "\\\\bdog".
}
