% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy-lazy-regex.R
\name{make_all_regex_greedy}
\alias{make_all_regex_greedy}
\title{Make regex patterns greedy}
\usage{
make_all_regex_greedy(pattern)
}
\arguments{
\item{pattern}{A regex pattern in lazy format}
}
\value{
The provided pattern in greedy format.
}
\description{
Returns the provided pattern in greedy format. The ? is removed
in all instances of ??, +?, *?, \{n,\}?, \{n,m\}?.
}
