% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy-lazy-regex.R
\name{make_all_regex_lazy}
\alias{make_all_regex_lazy}
\title{Make regex patterns lazy}
\usage{
make_all_regex_lazy(pattern)
}
\arguments{
\item{pattern}{A regex pattern in greedy format}
}
\value{
The provided pattern in lazy format.
}
\description{
Returns the provided pattern in lazy format. Repetition
operators are replaced by their lazy versions: ??, +?, *?, {n,}?, {n,m}?.
}
