% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-regex.R
\name{optimize_regex_patterns}
\alias{optimize_regex_patterns}
\title{Optimize regular expression}
\usage{
optimize_regex_patterns(patterns)
}
\arguments{
\item{patterns}{The regular expression which should be optimized.}
}
\value{
An optimized regular expression
}
\description{
Currently, word boundaries are added to regular expressions,
otherwise dictionary patterns would be open (e.g., "boy" would also match
"boycott"). If patterns start or end with a catchall wildcard "(.*)", this
expression is removed. These patterns are then processed much quicker.

This function allows to provide a list of patterns such as
c("girl", "boy(.\emph{)"). "girl" then only matches the world girl because it is
replaced with "\\bgirl\\b". "boy(.})" would match boy, boys, boycott, etc.
because it is replaced with "\\bboy" (which is also a lot quicker than
keeping "\\bboy(.*)").
}
