% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy-lazy-regex.R
\name{switch_regex_greedy_lazy}
\alias{switch_regex_greedy_lazy}
\title{Switch between lazy and greedy in regex patterns}
\usage{
switch_regex_greedy_lazy(pattern)
}
\arguments{
\item{pattern}{A regex pattern.}
}
\value{
The provided pattern with lazy operators switched to greedy and
greedy operators switched to lazy.
}
\description{
Lazy repetition operators are replaced by greedy ones and vice
versa.

On a minor note: Internally the pattern "\_LAAAAAZZZY170605\_" is used for
replacements. Thus, provided patterns should not include this pattern.
}
