% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-regex.R
\name{optimize_regex_patterns}
\alias{optimize_regex_patterns}
\title{Optimize regular expression}
\usage{
optimize_regex_patterns(patterns)
}
\arguments{
\item{patterns}{The regular expression which should be optimized.}
}
\value{
An optimized regular expression
}
\description{
Currently, word boundaries are added to regular expressions,
otherwise dictionary patterns would be open (e.g., "cat" would also match
"category"). If patterns start or end with a catchall wildcard "(.*)",
this expression is removed. These patterns are then processed much quicker.

This function allows to provide a list of patterns such as
c("dog", "cat(.*)"). "dog" then only matches the word dog because
it is replaced with "\\\\bdog\\\\b". "cat(.*)" would match cat, cats,
catastrophe, etc. (but not, e.g., tomcat) because it is replaced with
"\\\\bcat" (which is also a lot quicker than keeping
"\\\\bcat(.*)").

The function is intended to pre process patterns in the style of the
popdictR package.
}
